refactor(sw): extract the model-picker catalog to background/model-catalog.js
The SW is wiring, not logic (its own header rule), but the per-chat model
picker's whole catalog assembly lived inline: MODEL_CATALOG, the live
Ollama-inventory cache, the OpenRouter curated-selection mapping, the live
per-model context window, and buildModelOptions. Move it into an
import-free, deps-injected factory (makeModelCatalog) the same shape as
background/routes/*, so the assembly is Bun-testable with fakes and the SW
keeps only the one-line wiring.

localModelAvailable is passed as a thunk because localModelState is created
later in the SW's assembly order; the factory call sits after its
collaborators (safeFetch, getSecret, sessions) exist. The pricing↔catalog
parity test now reads the catalog literal from its new home. No behavior
change. ts-check floor 482 → 483.
