+++ tests/peerd-provider/pricing.test.ts
@@ -1,25 +1,25 @@
 // Pricing ↔ catalog parity guard.
 //
-// DEFAULT_PRICING (peerd-provider/pricing.js) and the SW's MODEL_CATALOG
-// drifted once already: catalog models with no rate card silently priced
-// at $0, and a rate card existed for a model that doesn't. The catalog
-// lives in background/service-worker.js, which can't be imported under
-// Bun (chrome.* at module scope) — so this test reads the SW SOURCE and
-// extracts the Anthropic catalog ids textually. If either table changes
-// without the other, this fails in the terminal.
+// DEFAULT_PRICING (peerd-provider/pricing.js) and MODEL_CATALOG drifted once
+// already: catalog models with no rate card silently priced at $0, and a rate
+// card existed for a model that doesn't. The catalog lives in
+// background/model-catalog.js, which can't be imported under Bun (it's assembled
+// against chrome.*-bound collaborators) — so this test reads the SOURCE and
+// extracts the Anthropic catalog ids textually. If either table changes without
+// the other, this fails in the terminal.
 
 import { describe, test, expect } from 'bun:test';
 import { readFileSync } from 'node:fs';
 import { join } from 'node:path';
 import { DEFAULT_PRICING, costOf, resolvePricing } from '../../extension/peerd-provider/pricing.js';
 
-const swPath = join(import.meta.dir, '../../extension/background/service-worker.js');
+const catalogPath = join(import.meta.dir, '../../extension/background/model-catalog.js');
 
-/** Extract the Anthropic model ids out of the SW's MODEL_CATALOG literal. */
+/** Extract the Anthropic model ids out of the MODEL_CATALOG literal. */
 const anthropicCatalogIds = (): string[] => {
-  const src = readFileSync(swPath, 'utf8');
+  const src = readFileSync(catalogPath, 'utf8');
   const catalog = src.match(/const MODEL_CATALOG = Object\.freeze\(\{([\s\S]*?)\}\);/);
-  if (!catalog) throw new Error('MODEL_CATALOG literal not found in service-worker.js');
+  if (!catalog) throw new Error('MODEL_CATALOG literal not found in model-catalog.js');
   const anthropic = catalog[1].match(/anthropic:\s*\[([\s\S]*?)\]/);
   if (!anthropic) throw new Error('anthropic entry not found in MODEL_CATALOG');
   return [...anthropic[1].matchAll(/model:\s*'([^']+)'/g)].map((m) => m[1]);
